Kopierat viss kod från Lorenzo ocg lagt till isActive i interfacet Student.
