+++ index.html
@@ -5,11 +5,12 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Studentkatalog</title>
     <link rel="stylesheet" href="style.css">
+    <script src="./dist/index.js" defer></script> 
 </head>
 <body>
     <h1>Studentkatalog</h1>
     <div id="student-list"></div>
     
-    <script type="module" src="src/main.js"></script>
+    <script type="module" src="dist/main.js" defer></script>
 </body>
 </html>
