--- package.json
@@ -1,24 +1,0 @@
-{
-    "name": "studentkatalog-agile-grupp1",
-    "version": "1.0.0",
-    "description": "En studentkatalog byggt med TypeScript för agile utveckling",
-    "main": "dist/main.js",
-    "scripts": {
-        "build": "tsc",
-        "dev": "tsc --watch",
-        "start": "npx http-server . -p 3000",
-        "clean": "rm -rf dist"
-    },
-    "keywords": [
-        "typescript",
-        "student",
-        "katalog",
-        "agile"
-    ],
-    "author": "",
-    "license": "MIT",
-    "devDependencies": {
-        "typescript": "^5.0.0",
-        "http-server": "^14.0.0"
-    }
-}
