+++ src/utils/studentModel.ts
@@ -1,0 +1,33 @@
+// Interface för student
+export interface Student {
+    id: number;
+    namn: string;
+    email: string;
+    kurs: string;
+    isActive: boolean;
+}
+
+// Hårdkodade studenter (minst 3 enligt US1)
+export const studenter: Student[] = [
+    {
+        id: 1,
+        namn: "Anna Andersson",
+        email: "anna.andersson@example.com",
+        kurs: "Webbutveckling",
+        isActive: true
+    },
+    {
+        id: 2,
+        namn: "Erik Eriksson",
+        email: "erik.eriksson@example.com",
+        kurs: "Frontend",
+        isActive: true
+    },
+    {
+        id: 3,
+        namn: "Sofia Nilsson",
+        email: "sofia.nilsson@example.com",
+        kurs: "Backend",
+        isActive: false
+    }
+];
